corrigindo advertencias do LivrosRepositorio

+++ ProjetoMinsait/Repositories/LivrosRepositorio.cs
@@ -22,6 +22,10 @@ namespace ProjetoMinsait.Repositories
         public async Task Delete(int id)
         {
             var livrosDelete = await _context.Livros.FindAsync(id);
+            if (livrosDelete == null)
+            {
+                throw new Exception($"Id: {id} não encontrado");
+            }
             _context.Livros.Remove(livrosDelete);
             await _context.SaveChangesAsync();
         }
@@ -33,7 +37,12 @@ namespace ProjetoMinsait.Repositories
 
         public async Task<Livros> Get(int id)
         {
-            return await _context.Livros.FindAsync(id);
+            var livrosProcurar = await _context.Livros.FindAsync(id);
+            if (livrosProcurar == null)
+            {
+                throw new Exception($"Id: {id} não encontrado");
+            }
+            return livrosProcurar;
         }
 
         public async Task Update(Livros livros)
